Chore : 프로젝트 재구조화
