+++ app/src/main/java/com/indoor/woopy_android/view/user/allfunction/AllFunctionFragment.kt
@@ -1,4 +1,4 @@
-package com.indoor.woopy_android.view.fragment
+package com.indoor.woopy_android.view.user.allfunction
 
 
 import android.content.Context
@@ -11,22 +11,22 @@ import androidx.fragment.app.Fragment
 import com.indoor.woopy_android.R
 
 
-class RequestFragment : Fragment() {
+class AllFunctionFragment : Fragment() {
 
     companion object{
         const val TAG: String = "로그"
-        fun newInstance() : RequestFragment{
-            return RequestFragment()
+        fun newInstance() : AllFunctionFragment {
+            return AllFunctionFragment()
         }
     }
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        Log.d(RequestFragment.TAG,"RequestFragment - onCreateView() called")
+        Log.d(TAG,"AllFunctionFragment - onCreate() called")
     }
     //프레그먼트를 안고 있는 엑티비티에 붙었을 때
     override fun onAttach(context: Context) {
         super.onAttach(context)
-        Log.d(RequestFragment.TAG,"RequestFragment - onCreateView() called")
+        Log.d(TAG,"AllFunctionFragment - onAttach() called")
     }
     //뷰가 생성되었을 때
     //프로그먼트와 레이아웃을 연결시켜주는 부분
@@ -35,9 +35,9 @@ class RequestFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        Log.d(RequestFragment.TAG,"RequestFragment - onCreateView() called")
+        Log.d(TAG,"AllFunctionFragment - onCreateView() called")
 
-        val view=inflater.inflate(R.layout.fragment_request,container,false)
+        val view=inflater.inflate(R.layout.fragment_all_function_fragment,container,false)
 
         return view
     }
