+++ app/src/main/java/com/indoor/woopy_android/view/user/approval/RequestFragment.kt
@@ -1,4 +1,4 @@
-package com.indoor.woopy_android.view.fragment
+package com.indoor.woopy_android.view.user.approval
 
 
 import android.content.Context
@@ -11,24 +11,22 @@ import androidx.fragment.app.Fragment
 import com.indoor.woopy_android.R
 
 
-class HomeFragment : Fragment() {
+class RequestFragment : Fragment() {
 
     companion object{
         const val TAG: String = "로그"
-        fun newInstance() : HomeFragment{
-            return HomeFragment()
+        fun newInstance() : RequestFragment {
+            return RequestFragment()
         }
     }
-
-    //메모리에 올라갔을 때
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        Log.d(TAG,"homeFragment - onCreate() called")
+        Log.d(TAG,"RequestFragment - onCreateView() called")
     }
     //프레그먼트를 안고 있는 엑티비티에 붙었을 때
     override fun onAttach(context: Context) {
         super.onAttach(context)
-        Log.d(TAG,"homeFragment - onAttach() called")
+        Log.d(TAG,"RequestFragment - onCreateView() called")
     }
     //뷰가 생성되었을 때
     //프로그먼트와 레이아웃을 연결시켜주는 부분
@@ -37,12 +35,10 @@ class HomeFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        Log.d(TAG,"homeFragment - onCreateView() called")
+        Log.d(TAG,"RequestFragment - onCreateView() called")
 
-        val view=inflater.inflate(R.layout.fragment_home,container,false)
+        val view=inflater.inflate(R.layout.fragment_request,container,false)
 
         return view
     }
-
-
 }
