+++ app/src/main/java/com/indoor/woopy_android/view/user/home/HomeFragment.kt
@@ -1,4 +1,4 @@
-package com.indoor.woopy_android.view.fragment
+package com.indoor.woopy_android.view.user.home
 
 
 import android.content.Context
@@ -11,22 +11,24 @@ import androidx.fragment.app.Fragment
 import com.indoor.woopy_android.R
 
 
-class AllFunctionFragment : Fragment() {
+class HomeFragment : Fragment() {
 
     companion object{
         const val TAG: String = "로그"
-        fun newInstance() : AllFunctionFragment{
-            return AllFunctionFragment()
+        fun newInstance() : HomeFragment {
+            return HomeFragment()
         }
     }
+
+    //메모리에 올라갔을 때
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        Log.d(AllFunctionFragment.TAG,"AllFunctionFragment - onCreate() called")
+        Log.d(TAG,"homeFragment - onCreate() called")
     }
     //프레그먼트를 안고 있는 엑티비티에 붙었을 때
     override fun onAttach(context: Context) {
         super.onAttach(context)
-        Log.d(AllFunctionFragment.TAG,"AllFunctionFragment - onAttach() called")
+        Log.d(TAG,"homeFragment - onAttach() called")
     }
     //뷰가 생성되었을 때
     //프로그먼트와 레이아웃을 연결시켜주는 부분
@@ -35,10 +37,12 @@ class AllFunctionFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        Log.d(AllFunctionFragment.TAG,"AllFunctionFragment - onCreateView() called")
+        Log.d(TAG,"homeFragment - onCreateView() called")
 
-        val view=inflater.inflate(R.layout.fragment_all_function_fragment,container,false)
+        val view=inflater.inflate(R.layout.fragment_home,container,false)
 
         return view
     }
+
+
 }
