+++ app/src/main/java/com/indoor/woopy_android/view/user/statistics/StatisticsFragment.kt
@@ -1,4 +1,4 @@
-package com.indoor.woopy_android.view.fragment
+package com.indoor.woopy_android.view.user.statistics
 
 
 import android.content.Context
@@ -15,18 +15,18 @@ class StatisticsFragment : Fragment() {
 
     companion object{
         const val TAG: String = "로그"
-        fun newInstance() : StatisticsFragment{
+        fun newInstance() : StatisticsFragment {
             return StatisticsFragment()
         }
     }
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        Log.d(StatisticsFragment.TAG,"StatisticsFragment - onCreate() called")
+        Log.d(TAG,"StatisticsFragment - onCreate() called")
     }
     //프레그먼트를 안고 있는 엑티비티에 붙었을 때
     override fun onAttach(context: Context) {
         super.onAttach(context)
-        Log.d(StatisticsFragment.TAG,"StatisticsFragment - onAttach() called")
+        Log.d(TAG,"StatisticsFragment - onAttach() called")
     }
     //뷰가 생성되었을 때
     //프로그먼트와 레이아웃을 연결시켜주는 부분
@@ -35,7 +35,7 @@ class StatisticsFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        Log.d(StatisticsFragment.TAG,"StatisticsFragment - onCreateView() called")
+        Log.d(TAG,"StatisticsFragment - onCreateView() called")
 
         val view=inflater.inflate(R.layout.fragment_statistics,container,false)
