+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".view.activity.MainActivity">
+        tools:context=".view.user.MainActivity">
 
         <FrameLayout
             android:id="@+id/fragments_frame"
